+++ gdsfactory/components/straight_heater_doped_rib.py
@@ -195,12 +195,11 @@ def straight_heater_doped_rib(
     return c
 
 
-def test_straight_heater_doped_rib_ports() -> Component:
+def test_straight_heater_doped_rib_ports() -> None:
     c = straight_heater_doped_rib(length=100.0)
     assert c.get_ports_xsize(port_type="optical") == 100.0, c.get_ports_xsize(
         port_type="optical"
     )
-    return c
 
 
 if __name__ == "__main__":
