+++ test-data-regression/test_settings_pad_array270_.yml
@@ -8,7 +8,7 @@ ports:
     - 49
     - 0
     name: e11
-    orientation: 270
+    orientation: 270.0
     port_type: electrical
     shear_angle: null
     width: 100.0
@@ -20,7 +20,7 @@ ports:
     - 49
     - 0
     name: e12
-    orientation: 270
+    orientation: 270.0
     port_type: electrical
     shear_angle: null
     width: 100.0
@@ -32,7 +32,7 @@ ports:
     - 49
     - 0
     name: e13
-    orientation: 270
+    orientation: 270.0
     port_type: electrical
     shear_angle: null
     width: 100.0
@@ -44,7 +44,7 @@ ports:
     - 49
     - 0
     name: e14
-    orientation: 270
+    orientation: 270.0
     port_type: electrical
     shear_angle: null
     width: 100.0
@@ -56,7 +56,7 @@ ports:
     - 49
     - 0
     name: e15
-    orientation: 270
+    orientation: 270.0
     port_type: electrical
     shear_angle: null
     width: 100.0
@@ -68,7 +68,7 @@ ports:
     - 49
     - 0
     name: e16
-    orientation: 270
+    orientation: 270.0
     port_type: electrical
     shear_angle: null
     width: 100.0
