+++ test-data-regression/test_settings_ring_double_heater_.yml
@@ -98,7 +98,7 @@ ports:
     width: 0.5
   r_e1:
     center:
-    - 0.0
+    - -0.0
     - 0.0
     layer:
     - 49
