+++ tests/routing/test_routing_get_route_error.py
@@ -6,7 +6,7 @@ import gdsfactory as gf
 from gdsfactory.routing.manhattan import RouteWarning
 
 
-def test_route_error():
+def test_route_error() -> None:
     """Ensures that an impossible route raises RouteWarning."""
     c = gf.Component("test_route_error")
     w = gf.components.straight()
@@ -30,9 +30,26 @@ def test_route_error():
         )
         c.add(route.references)
         c.add(route.labels)
-    return c
+
+
+def test_route_error2() -> None:
+    """Impossible route."""
+    c = gf.Component("pads_route_from_steps")
+    pt = c << gf.components.pad_array(orientation=270, columns=3)
+    pb = c << gf.components.pad_array(orientation=90, columns=3)
+    pt.move((100, 200))
+    route = gf.routing.get_route_from_steps(
+        pt.ports["e11"],
+        pb.ports["e11"],
+        steps=[
+            {"y": 100},
+        ],
+        cross_section="metal_routing",
+        bend=gf.components.wire_corner,
+    )
+    c.add(route.references)
+    c.add(route.labels)
 
 
 if __name__ == "__main__":
-    c = test_route_error()
-    c.show(show_ports=True)
+    test_route_error()
